+++ assignment_5/DatabaseRoutine/src/main/java/it/unitn/disi/webarch/mstolin/entities/accommodation/HotelEntity.java
@@ -8,7 +8,7 @@ public class HotelEntity extends AccommodationEntity {
 
     @Basic
     @Column(name = "EXTRA_HALF_BOARD", nullable = true)
-    private Integer extraHalfBoard;
+    private Double extraHalfBoard;
 
     @Basic
     @Column(name = "STARS", nullable = true)
@@ -18,7 +18,7 @@ public class HotelEntity extends AccommodationEntity {
     @Column(name = "PLACES", nullable = true)
     private Integer places;
 
-    public HotelEntity(String name, int price, int extraHalfBoard, int stars, int places) {
+    public HotelEntity(String name, Double price, Double extraHalfBoard, int stars, int places) {
         super();
         this.name = name;
         this.price = price;
@@ -31,11 +31,11 @@ public class HotelEntity extends AccommodationEntity {
         super();
     }
 
-    public Integer getExtraHalfBoard() {
+    public Double getExtraHalfBoard() {
         return extraHalfBoard;
     }
 
-    public void setExtraHalfBoard(Integer extraHalfBoard) {
+    public void setExtraHalfBoard(Double extraHalfBoard) {
         this.extraHalfBoard = extraHalfBoard;
     }
